.history/requirements_20251209004751.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13 
Pillow>=8.2.0,<8.3
psycopg2-binary>=2.8.6,<2.9